feat: fixed file locations, add a  json in assets to mock info in categories, add a component called CategoriesCard to implement in app with the respective css, changes in app to receive info in card, changes in app.css to organize the cards an add a title of cards
