+++ frontend/reactapp/src/App.css
@@ -13,6 +13,7 @@
   padding: 2rem 8rem 0 0;
   font-family: Arial, Helvetica, sans-serif monospace;
   font-size: 6rem;
+  color: #404145;
   background-color: white;
 }
 .containerCard{
