+++ frontend/reactapp/src/Components/CategoriesCard/categoriesCard.css
@@ -16,7 +16,7 @@
     width: 273px;
     border-radius: 1rem;
     padding: 1rem;
-    font-family:Arial, Helvetica, sans-serif monospace;
+    font-family:Helvetica, sans-serif;
     color: white;
     background:linear-gradient(rgba(0, 0, 0, 0.753), rgba(128, 128, 128, 0))
