+++ frontend/reactapp/src/Components/HomeBrand/Homebrand.jsx
@@ -1,0 +1,11 @@
+import './HomeBrand.css'
+
+function HomeBrand(){
+    return(
+        <section className='brand'>
+            <img className="logo" src="https://raw.githubusercontent.com/santy289/WorkIt/19030bfe4051391a4dd9b83aa12369eb18f2a1e2/frontend/reactapp/src/assets/Logoworkit.svg" alt="logo" />
+        </section>
+    )
+}
+
+export default HomeBrand;
