--- frontend/reactapp/src/Components/Homebrand.jsx
@@ -1,12 +1,0 @@
-import image from '../assets/logo.png'
-import './HomeBrand.css'
-
-function HomeBrand(){
-    return(
-        <section className='brand'>
-            <img src={image} className="logo" alt="logo" />
-        </section>
-    )
-}
-
-export default HomeBrand;
